+++ package.json
@@ -22,6 +22,7 @@
     "prism-react-renderer": "^2.3.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.9.6",
     "react-type-animation": "^3.2.0"
   },
   "devDependencies": {
