+++ src/components/menteeCard.jsx
@@ -12,7 +12,7 @@ const MenteeCard = ({ mentee }) => {
       <div className="p-6 bg-[#ac6dff]">
         <h3 className="text-xl font-bold text-white">{mentee.name}</h3>
         <p className="text-slate-300 mt-1">{mentee.role}</p>
-        <Link className="bg-white no-underline px-4 py-2 rounded-md" to="/cohorts/api/first/user">View projects</Link>
+        <Link className="bg-white no-underline px-4 py-2 rounded-md" to={`/cohorts/api/mentee/${mentee.slug}`}>View projects</Link>
       </div>
     </div>
   );
