+++ src/pages/cohorts/api/menteesDetails/first/mentees.json
@@ -1,9 +1,10 @@
 [
     {
         "name": "Edun Rilwan",
+        "slug": "edun-rilwan",
         "role": "Technical writer & Software developer",
         "bio": "I build Software solutions and write about technology",
-        "imgURL": "https://media.licdn.com/dms/image/v2/D5603AQGtHMpTYdjWwA/profile-displayphoto-shrink_400_400/B56ZTRxPe1HwAg-/0/1738686135680?e=1757548800&v=beta&t=tBNXXL-tZJpxY4qpZ05BeLNd7Gcd-JiRr-NlGcrg8Bo",
+        "imgURL": "https://media.licdn.com/dms/image/v2/D5603AQGtHMpTYdjWwA/profile-displayphoto-shrink_400_400/B56ZTRxPe1HwAg-/0/1738686135680?e=1764806400&v=beta&t=QRvLYprq1Z7XjgURLyhxoTuXBdtsxKMScYxt-3kxfyo",
         "projects": [
             {
                 "type": "api",
@@ -25,9 +26,10 @@
 
     {
         "name": "Etugbo Judith",
+        "slug": "etugbo-judith",
         "role": "Technical writer & Software developer",
         "bio": "I build Software solutions and write about technology",
-        "imgURL": "https://media.licdn.com/dms/image/v2/D4D03AQFGtXJS1tDy0Q/profile-displayphoto-scale_400_400/B4DZhhb.HUHwAg-/0/1753981381054?e=1757548800&v=beta&t=wvSJwVICS8XnS4zm7FEK1BQieye5b21nqUgCg07DXBE",
+        "imgURL": "https://media.licdn.com/dms/image/v2/D4D03AQFZOvk_Ob2aoA/profile-displayphoto-scale_400_400/B4DZpE4I0EJMAg-/0/1762092162697?e=1764806400&v=beta&t=FSxnOvceOb8SwClz4r6mybnDh1vudXM7pGS32XGrPnU",
         "projects": [
             {
                 "type": "api",
@@ -48,9 +50,10 @@
 
     {
         "name": "Peace Sandy",
+        "slug": "peace-sandy",
         "role": "Technical writer & Software developer",
         "bio": "I build Software solutions and write about technology",
-        "imgURL": "https://media.licdn.com/dms/image/v2/D4D03AQF48XBl0sVXkA/profile-displayphoto-scale_400_400/B4DZfKw0CXHkAk-/0/1751453486115?e=1757548800&v=beta&t=OgFLOAbtp9rYfTv-LB49UIJMyTTZbPfR9FJDWBpjSNI",
+        "imgURL": "https://media.licdn.com/dms/image/v2/D4D03AQFc16qt7qWnEw/profile-displayphoto-scale_400_400/B4DZjbzmUiGQAg-/0/1756034395136?e=1764806400&v=beta&t=VHHFdFzZdjMqIUiKj8_Fdd1QiYi38kOl0wzyRR3FcL8",
         "projects": [
             {
                 "type": "api",
